chore: extract parseAnalysisResults to a new module
clean up index.ts a bit
